Moving GithubCommit to use yaml config. Working on #78
