+++ spec/DevBoard/Github/Branch/Entity/GithubBranchSpec.php
@@ -1,7 +1,7 @@
 <?php
 namespace spec\DevBoard\Github\Branch\Entity;
 
-use DevBoard\Github\Commit\GithubCommit;
+use DevBoard\Github\Commit\Entity\GithubCommit;
 use DevBoard\Github\Repo\Entity\GithubRepo;
 use PhpSpec\ObjectBehavior;
 use Prophecy\Argument;
