+++ spec/DevBoard/Github/Commit/Entity/GithubCommitRepositorySpec.php
@@ -1,5 +1,5 @@
 <?php
-namespace spec\DevBoard\Github\Commit;
+namespace spec\DevBoard\Github\Commit\Entity;
 
 use Doctrine\ORM\EntityManager;
 use Doctrine\ORM\Mapping\ClassMetadata;
@@ -10,7 +10,7 @@ class GithubCommitRepositorySpec extends ObjectBehavior
 {
     public function it_is_initializable()
     {
-        $this->shouldHaveType('DevBoard\Github\Commit\GithubCommitRepository');
+        $this->shouldHaveType('DevBoard\Github\Commit\Entity\GithubCommitRepository');
     }
 
     public function let(EntityManager $em, ClassMetadata $classMetadata)
