+++ src/DevBoard/Behat/Github/Commit/FixtureContext.php
@@ -2,7 +2,7 @@
 namespace DevBoard\Behat\Github\Commit;
 
 use Behat\Gherkin\Node\TableNode;
-use DevBoard\Github\Commit\GithubCommit;
+use DevBoard\Github\Commit\Entity\GithubCommit;
 use Resources\Behat\DefaultContext;
 
 /**
