+++ src/DevBoard/Github/Branch/Entity/GithubBranch.php
@@ -1,7 +1,7 @@
 <?php
 namespace DevBoard\Github\Branch\Entity;
 
-use DevBoard\Github\Commit\GithubCommit;
+use DevBoard\Github\Commit\Entity\GithubCommit;
 use DevBoard\Github\Repo\Entity\GithubRepo;
 use Doctrine\ORM\Mapping as ORM;
 use Resources\Entity\BaseEntity;
